requirements.txt
# Core Comparison Frameworks
customtkinter
flet
nicegui
PySide6
PyQt6
wxPython
PySimpleGUI
dearpygui
kivy
toga
eel
pywebview
flexx
appJar
gooey

# Additional dependencies
# Note: datetime is part of Python standard library, no installation needed
# requests: Currently not used, but kept for potential future use
